Split tests/api/api.spec.ts
